+++ src/Formatting.Analyzers/CSharp/NormalizeWhitespaceAtEndOfFileAnalyzer.cs
@@ -83,10 +83,16 @@ public sealed class NormalizeWhitespaceAtEndOfFileAnalyzer : BaseDiagnosticAnaly
         }
         else if (en.MoveNext())
         {
-            if (CSharpFacts.IsCommentTrivia(en.Current.Kind())
-                || SyntaxFacts.IsPreprocessorDirective(en.Current.Kind()))
+            if (CSharpFacts.IsCommentTrivia(en.Current.Kind()))
             {
                 ReportDiagnostic(context, endOfFile);
+            }
+            else if (SyntaxFacts.IsPreprocessorDirective(en.Current.Kind()))
+            {
+                var directive = (DirectiveTriviaSyntax)en.Current.GetStructure();
+
+                if (!directive.GetTrailingTrivia().LastOrDefault().IsKind(SyntaxKind.EndOfLineTrivia))
+                    ReportDiagnostic(context, endOfFile);
             }
             else if (en.Current.IsWhitespaceOrEndOfLineTrivia()
                 && endOfFile.LeadingTrivia.Span.Start == 0)